signup converted to modal, very basic styling added
