+++ frontend/src/components/LoginFormModal/index.js
@@ -7,9 +7,10 @@ function LoginFormModal() {
 
   return (
     <>
-      <button onClick={() => setShowModal(true)}>Log In</button>
+      <button 
+      id='login-button' onClick={() => setShowModal(true)}>Log In</button>
       {showModal && (
-        <Modal onClose={() => setShowModal(false)}>
+        <Modal className="login-modal" onClose={() => setShowModal(false)}>
           <LoginForm />
         </Modal>
       )}
