+++ frontend/src/components/Navigation/Navigation.css
@@ -37,7 +37,6 @@
   list-style: none;
   position: fixed;
   background-color: white;
-  border: 1px solid #ff385c;
   gap: 3px;
   justify-content: center;
   align-items: center;
