+++ frontend/src/components/SignupFormPage/SignupForm.css
@@ -1,7 +1,12 @@
-.signup-form-container {
+.signup-form {
+  width: 66%;
   display: flex;
   flex-direction: column;
-  justify-content: center;
+  padding: 2%;
+  justify-content: space-between;
   align-items: center;
   flex-wrap: wrap;
+  box-sizing: border-box;
 }
+
+
