+++ frontend/src/index.css
@@ -1,1 +1,8 @@
 /* TODO Add site wide styles */
+body {
+  font-family: sans-serif;
+}
+
+button:hover {
+  background-color: #f7f7f7;
+}
